refactor(audio): move SYNTH_PRESETS to tone-free chatSynthPresets module
